Move {Browser,Jsdom}PageOpener to separate modules
This makes things easier to read, and will make testing some details of
the separate implementations more straightforward.

Also created the new lib/types.js to share Jsdoc typedefs based on a
hint from: https://stackoverflow.com/a/76872194
